1) change timer field from Dict[str, Interval] to Dict[str, List[Interval]] 2) make timer available to all entities via ReportLink linkage to parent's report 3) remove EventRecorder 4) re-org Report class hierarchy 5) re-org ReportSchema class hierarchy 6) UI changes and tests
