+++ doc/newsfragments/2668_changed.new_timer.rst
@@ -1,0 +1,2 @@
+Enhance timer field of testplan report to allow multiple records under the same key. We now merge multitest parts aggregating their runtime. Also refactored Report and ReportSchema class hierachy.
+
