+++ testplan/common/entity/base.py
@@ -27,13 +27,23 @@ import psutil
 from schema import Or
 
 from testplan.common.config import Config, ConfigOption
-from testplan.common.report.base import EventRecorder
 from testplan.common.utils import logger
 from testplan.common.utils.path import default_runpath, makedirs, makeemptydirs
 from testplan.common.utils.strings import slugify, uuid4
 from testplan.common.utils.thread import execute_as_thread, interruptible_join
-from testplan.common.utils.timing import wait
+from testplan.common.utils.timing import wait, Timer
 from testplan.common.utils.validation import is_subclass
+
+
+class ReportLink:
+    """
+    A recursive linkage that will be available to all Entity object.
+    """
+
+    def __init__(self, name: str):
+        self.name = name
+        self.timer = Timer()
+        self.children: List["ReportLink"] = []
 
 
 class Environment:
@@ -523,6 +533,7 @@ class Entity(logger.Loggable):
         self._uid = None
         self._should_abort = False
         self._aborted = False
+        self._report = None
 
     def __str__(self):
         return f"{self.__class__.__name__}[{self.uid()}]"
@@ -582,6 +593,28 @@ class Entity(logger.Loggable):
         Reference to parent object.
         """
         self._parent = value
+
+    @property
+    def report(self):
+        """
+        A handle to access the report via recursive parent
+        """
+        if not self._report:
+            self._report = ReportLink(name=str(self))
+            if self.parent and hasattr(self.parent, "report"):
+                # don't always have parent in test
+                # ChildLoop doesn't have report
+                self.parent.report.children.append(self._report)
+            else:
+                # will just have a dangling child
+                # but the timer won't be collected to report
+                self.logger.debug("dangling report")
+
+        return self._report
+
+    @property
+    def timer(self):
+        return self.report.timer
 
     def pause(self):
         """
@@ -863,17 +896,9 @@ class Runnable(Entity):
         self._environment: Environment = self.__class__.ENVIRONMENT(
             parent=self
         )
-        self._result: RunnableResult = self.__class__.RESULT()
+        self.result: RunnableResult = self.__class__.RESULT()
         self._steps: Deque[Tuple[Callable, List, Dict]] = deque()
         self._ihandler = None
-
-    @property
-    def result(self):
-        """
-        Returns a
-        :py:class:`~testplan.common.entity.base.RunnableResult`
-        """
-        return self._result
 
     @property
     def resources(self):
@@ -1190,15 +1215,15 @@ class Runnable(Entity):
             else:
                 self._run_batch_steps()
         except Exception as exc:
-            self._result.run = exc
+            self.result.run = exc
             self.logger.error(traceback.format_exc())
         else:
-            # TODO fix swallow exceptions in self._result.step_results.values()
-            self._result.run = (
+            # TODO fix swallow exceptions in self.result.step_results.values()
+            self.result.run = (
                 self.status == RunnableStatus.FINISHED
                 and self.run_result() is True
             )
-        return self._result
+        return self.result
 
     def run_result(self):
         """
@@ -1206,7 +1231,7 @@ class Runnable(Entity):
         """
         return all(
             not isinstance(val, Exception) and val is not False
-            for val in self._result.step_results.values()
+            for val in self.result.step_results.values()
         )
 
     def dry_run(self):
@@ -1319,9 +1344,6 @@ class Resource(Entity):
                 self.STATUS.STOPPED: self._wait_stopped,
             }
         )
-        self.event_recorder: EventRecorder = EventRecorder(
-            name=self.uid(), event_type="Executor"
-        )
 
     @property
     def context(self):
@@ -1373,7 +1395,7 @@ class Resource(Entity):
             )
             return
 
-        self.event_recorder.start_time = time.time()
+        self.timer.start("lifespan")
 
         self.logger.info("Starting %s", self)
         self.status.change(self.STATUS.STARTING)
@@ -1419,7 +1441,7 @@ class Resource(Entity):
         if not self.async_start:
             self.wait(self.STATUS.STOPPED)
             self.logger.info("%s stopped", self)
-        self.event_recorder.end_time = time.time()
+        self.timer.end("lifespan")
 
     def pre_start(self):
         """
@@ -1659,6 +1681,10 @@ class RunnableManager(Entity):
         """
         runnable_class = self._cfg.runnable
         return runnable_class(**options)
+
+    @property
+    def report(self):
+        return self.runnable.report
 
     def __getattr__(self, item):
         try:
