+++ testplan/report/__init__.py
@@ -3,8 +3,10 @@ from .testing import (
     TestReport,
     TestGroupReport,
     TestCaseReport,
+    styles as test_styles,
+)
+from testplan.common.report import (
     Status,
     RuntimeStatus,
-    styles as test_styles,
     ReportCategories,
 )
