+++ testplan/report/testing/__init__.py
@@ -2,9 +2,6 @@
 
 from . import styles
 from .base import (
-    ReportCategories,
-    RuntimeStatus,
-    Status,
     TestCaseReport,
     TestGroupReport,
     TestReport,
