+++ testplan/report/testing/base.py
@@ -45,507 +45,19 @@ import itertools
 import os
 import platform
 import sys
-import traceback
 from collections import Counter
-from enum import Enum, auto
-from functools import reduce, total_ordering
-from typing import Callable, Dict, List, Optional
-
+from typing import Dict, Optional
 from typing_extensions import Self
 
-from testplan.common.report import ExceptionLogger as ExceptionLoggerBase
-from testplan.common.report import Report, ReportGroup, SkipTestcaseException
-from testplan.common.utils import timing
+from testplan.common.report import (
+    Status,
+    RuntimeStatus,
+    ReportCategories,
+    ExceptionLogger,
+    Report,
+    BaseReportGroup,
+)
 from testplan.testing import tagging
-
-
-@total_ordering
-class RuntimeStatus(Enum):
-    """
-    Constants for test runtime status - for interactive mode
-
-    total order encoded in value, serialized to "lower-case string" of name
-    """
-
-    RUNNING = auto()
-    RESETTING = auto()
-    WAITING = auto()
-    READY = auto()
-    NOT_RUN = auto()
-    FINISHED = auto()
-    NONE = auto()  # an "unset" value
-
-    @classmethod
-    def precedent(cls, stats: List[Self]) -> Self:
-        """
-        Return the precedent status from a list of statuses, using the
-        ordering of statuses in `rule`.
-
-        Note that the client can send RESETTING signal to reset the test report
-        to its initial status, but client will not receive a temporary report
-        containing RESETTING status, instead WAITING status is used and after
-        reset the report goes to READY status.
-
-        :param stats: List of statuses of which we want to get the precedent.
-        """
-        return min(stats, key=lambda x: x.value)
-
-    def __lt__(self, other: Self) -> bool:
-        return self.value < other.value
-
-    precede = __lt__
-
-    def __bool__(self):
-        return self != self.NONE
-
-    def to_json_compatible(self) -> Optional[str]:
-        if self.name == "NONE":
-            return None
-        return self.name.lower()
-
-    @classmethod
-    def from_json_compatible(cls, s: Optional[str]) -> Self:
-        if s is None:
-            return cls.NONE
-        return cls[s.upper()]
-
-
-class Status(Enum):
-    """
-    Constants for test result and utilities for propagating status upward.
-
-    partial order encoded by value, serialized to "lower-case string" of name
-
-    tens of value encoding status category, tenths of value only for
-    differentiating enum members
-    """
-
-    BOTTOM = -1  # minimal
-    ERROR = 9
-    INCOMPLETE = 18.1
-    XPASS_STRICT = 18.2
-    FAILED = 19  # red
-    UNKNOWN = 29  # black
-    PASSED = 39  # green
-    SKIPPED = 48.1
-    XFAIL = 48.2
-    XPASS = 48.3
-    UNSTABLE = 49  # yellow
-    NONE = 59  # maxium, "unset"
-
-    @classmethod
-    def precedent(cls, stats: List[Self]) -> Self:
-        """
-        Return the precedent status from a list of statuses, using the
-        ordering of statuses in `rule`.
-
-        :param stats: List of statuses of which we want to get the precedent.
-        """
-
-        # unrelated pair fallback to norm
-        def _cmp(x: Self, y: Self) -> Self:
-            try:
-                r = x < y
-            except TypeError:
-                return x.normalised()
-            else:
-                return x if r else y
-
-        return reduce(_cmp, stats, cls.NONE)
-
-    def __lt__(self, other: Self) -> bool:
-        lhs, rhs = int(self.value), int(other.value)
-        if lhs == rhs and self != other:
-            return NotImplemented
-        return lhs < rhs
-
-    def __le__(self, other: Self) -> bool:
-        lhs, rhs = int(self.value), int(other.value)
-        if lhs == rhs and self != other:
-            return NotImplemented
-        return lhs <= rhs
-
-    def precede(self, other: Self) -> bool:
-        # a more intuitive & exception-free version
-        try:
-            return self < other
-        except TypeError:
-            return False
-
-    def normalised(self) -> Self:
-        return self.__class__(self.value // 10 * 10 + 9)
-
-    def __bool__(self) -> bool:
-        return self != self.NONE and self != self.BOTTOM
-
-    def to_json_compatible(self) -> Optional[str]:
-        if self.name == "NONE":
-            return None
-        return self.name.lower().replace("_", "-")
-
-    @classmethod
-    def from_json_compatible(cls, s: Optional[str]) -> Self:
-        if s is None:
-            return cls.NONE
-        return cls[s.replace("-", "_").upper()]
-
-
-class ReportCategories:
-    """
-    Enumeration of possible categories of report nodes.
-
-    Note: we don't use the enum.Enum base class to simplify report
-    serialization via marshmallow.
-    """
-
-    TESTPLAN = "testplan"
-    TESTGROUP = "testgroup"  # test group of unspecific type?
-    MULTITEST = "multitest"
-    TASK_RERUN = "task_rerun"
-    TESTSUITE = "testsuite"
-    TESTCASE = "testcase"
-    PARAMETRIZATION = "parametrization"
-    GTEST = "gtest"
-    GTEST_SUITE = "gtest-suite"
-    CPPUNIT = "cppunit"
-    CPPUNIT_SUITE = "cppunit-suite"
-    BOOST_TEST = "boost-test"
-    BOOST_SUITE = "boost-suite"
-    HOBBESTEST = "hobbestest"
-    HOBBESTEST_SUITE = "hobbestest-suite"
-    PYTEST = "pytest"
-    PYUNIT = "pyunit"
-    UNITTEST = "unittest"
-    QUNIT = "qunit"
-    JUNIT = "junit"
-    ERROR = "error"
-    # use for before/after_start/stop, setup, teardown, etc
-    SYNTHESIZED = "synthesized"
-
-
-class ExceptionLogger(ExceptionLoggerBase):
-    """
-    When we run tests, we always want to return a report object,
-    However we also want to mark the test as failed if an
-    exception is raised (unless kwargs['fail'] is `False`).
-    """
-
-    def __init__(self, *exception_classes, **kwargs):
-        self.fail = kwargs.get("fail", True)
-        super(ExceptionLogger, self).__init__(*exception_classes, **kwargs)
-
-    def __exit__(self, exc_type, exc_value, tb):
-        if exc_type is not None:
-            if exc_type is SkipTestcaseException:
-                self.report.logger.critical(
-                    'Skipping testcase "%s", reason: %s',
-                    self.report.name,
-                    str(exc_value),
-                )
-                self.report.status_override = Status.SKIPPED
-            elif issubclass(exc_type, self.exception_classes):
-                # Custom exception message with extra args
-                exc_msg = "".join(
-                    traceback.format_exception(exc_type, exc_value, tb)
-                )
-                self.report.logger.error(exc_msg)
-
-                if self.fail:
-                    self.report.status_override = Status.ERROR
-            return True
-
-
-class BaseReportGroup(ReportGroup):
-    """Base container report for tests, relies on children's statuses."""
-
-    exception_logger = ExceptionLogger
-
-    def __init__(self, name, **kwargs):
-        self.meta = kwargs.pop("meta", {})
-        self.status_override = kwargs.pop("status_override", Status.NONE)
-        self.status_reason = kwargs.pop("status_reason", None)
-
-        super(BaseReportGroup, self).__init__(name=name, **kwargs)
-
-        self.timer = timing.Timer()
-
-        # Normally, a report group inherits its statuses from its child
-        # entries. However, in case there are no child entries we use the
-        # following values as a fallback.
-        self._status = Status.UNKNOWN
-        self._runtime_status = RuntimeStatus.READY
-
-    def _get_comparison_attrs(self):
-        return super(BaseReportGroup, self)._get_comparison_attrs() + [
-            "status_override",
-            "timer",
-        ]
-
-    @property
-    def passed(self):
-        """Shortcut for getting if report status should be considered passed."""
-        return self.status.normalised() == Status.PASSED
-
-    @property
-    def failed(self):
-        """
-        Shortcut for checking if report status should be considered failed.
-        """
-        return self.status <= Status.FAILED
-
-    @property
-    def unstable(self):
-        """
-        Shortcut for checking if report status should be considered unstable.
-        """
-        return self.status.normalised() == Status.UNSTABLE
-
-    @property
-    def unknown(self):
-        """
-        Shortcut for checking if report status is unknown.
-        """
-        return self.status.normalised() == Status.UNKNOWN
-
-    @property
-    def status(self):
-        """
-        Status of the report, will be used to decide
-        if a Testplan run has completed successfully or not.
-
-        `status_override` always takes precedence,
-        otherwise we fall back to precedent status from `self.entries`.
-
-        If a report group has no children, it is assumed to be passing.
-        """
-        if self.status_override:
-            return self.status_override
-
-        if self.entries:
-            return Status.precedent([entry.status for entry in self])
-
-        return self._status
-
-    @status.setter
-    def status(self, new_status):
-        self._status = new_status
-
-    @property
-    def runtime_status(self):
-        """
-        The runtime status is used for interactive running, and reports
-        whether a particular entry is READY, WAITING, RUNNING, RESETTING,
-        FINISHED or NOT_RUN.
-
-        A test group inherits its runtime status from its child entries.
-        """
-        if self.entries:
-            return RuntimeStatus.precedent(
-                [entry.runtime_status for entry in self]
-            )
-
-        return self._runtime_status
-
-    @runtime_status.setter
-    def runtime_status(self, new_status):
-        """Set the runtime_status of all child entries."""
-        for entry in self:
-            # TODO: use suite_related flag for now, use synthesized instead
-            if not getattr(entry, "suite_related", False):
-                entry.runtime_status = new_status
-        self._runtime_status = new_status
-
-    def set_runtime_status_filtered(
-        self,
-        new_status: str,
-        entries: Dict,
-    ) -> None:
-        """
-        Alternative setter for the runtime status of an entry. Propagates only
-          to the specified entries.
-
-        :param new_status: new runtime status to be set
-        :param entries: tree-like structure of entries names
-        """
-        for entry in self:
-            if entry.name in entries.keys():
-                if isinstance(entry, TestCaseReport):
-                    entry.runtime_status = new_status
-                else:
-                    entry.set_runtime_status_filtered(
-                        new_status, entries[entry.name]
-                    )
-        self._runtime_status = new_status
-
-    def merge_children(self, report, strict=True):
-        """
-        For report groups, we call `merge` on each child report
-        and later merge basic attributes.
-
-        However sometimes original report may not have matching child reports.
-        For example this happens when the test target's test cases cannot be
-        inspected on initialization time (e.g. GTest).
-
-        In this case we can merge with `strict=False` so that child reports
-        are directly appended to original report's entries.
-        """
-        if strict:
-            super(BaseReportGroup, self).merge_children(report, strict=strict)
-        else:
-            for entry in report:
-                if entry.uid not in self._index:
-                    self.append(entry)
-                else:
-                    self.get_by_uid(entry.uid).merge(entry, strict=strict)
-
-    def merge(self, report, strict=True):
-        """Update `status_override` as well."""
-        super(BaseReportGroup, self).merge(report, strict=strict)
-
-        self.timer.update(report.timer)
-        self.status_override = Status.precedent(
-            [self.status_override, report.status_override],
-        )
-
-    @property
-    def counter(self) -> Counter:
-        """
-        Return counts for each status, will recursively get aggregates from
-        children and so on.
-        """
-        counter = Counter(
-            {
-                Status.PASSED.to_json_compatible(): 0,
-                Status.FAILED.to_json_compatible(): 0,
-                "total": 0,
-            }
-        )
-
-        # exclude rerun and synthesized entries from counter
-        for child in self:
-            if child.category == ReportCategories.ERROR:
-                counter.update(
-                    {Status.ERROR.to_json_compatible(): 1, "total": 1}
-                )
-            elif child.category in (
-                ReportCategories.TASK_RERUN,
-                ReportCategories.SYNTHESIZED,
-            ):
-                pass
-            else:
-                counter.update(child.counter)
-
-        return counter
-
-    def filter(self, *functions, **kwargs) -> Self:
-        """
-        Tag indices are updated after filter operations.
-        """
-        result = super(BaseReportGroup, self).filter(*functions, **kwargs)
-
-        # We'd like to call tag propagation before returning the root node,
-        # so we rely on absence of implicit `__copy` arg to decide if we should
-        # trigger tag index propagation or not. If we don't do this check
-        # then tag propagation will be called for each filter call on
-        # sub-nodes which is going to be a redundant operation.
-
-        if kwargs.get("__copy", True):
-            result.propagate_tag_indices()
-        return result
-
-    def filter_cases(
-        self, predicate: Callable[["ReportGroup"], bool], is_root: bool = False
-    ) -> Self:
-        """
-        Case-level filter with status retained
-        """
-        report_obj = copy.deepcopy(self) if is_root else self
-        statuses = []
-        entries = []
-
-        for entry in report_obj.entries:
-            if isinstance(entry, BaseReportGroup):
-                statuses.append(entry.status)
-                entry = entry.filter_cases(predicate)
-                if len(entry):
-                    entries.append(entry)
-            elif isinstance(entry, TestCaseReport):
-                statuses.append(entry.status)
-                if predicate(entry):
-                    entries.append(entry)
-            else:
-                raise TypeError(
-                    f"Unexpected entry {entry} of type {type(entry)} here."
-                )
-
-        report_obj.entries = entries
-        report_obj.status_override = Status.precedent(statuses)
-        if is_root:
-            report_obj.build_index(recursive=True)
-
-        return report_obj
-
-    def filter_by_tags(self, tag_value, all_tags=False) -> Self:
-        """Shortcut method for filtering the report by given tags."""
-
-        def _filter_func(obj):
-            # Include all testcase entries, which are in dict form
-            if isinstance(obj, dict):
-                return True
-
-            tag_dict = tagging.validate_tag_value(tag_value)
-            if all_tags:
-                match_func = tagging.check_all_matching_tags
-            else:
-                match_func = tagging.check_any_matching_tags
-
-            return match_func(
-                tag_arg_dict=tag_dict, target_tag_dict=obj.tags_index
-            )
-
-        return self.filter(_filter_func)
-
-    @property
-    def hash(self) -> int:
-        """
-        Generate a hash of this report object, including its entries. This
-        hash is used to detect when changes are made under particular nodes
-        in the report tree. Since all report entries are mutable, this hash
-        should NOT be used to index the report entry in a set or dict - we
-        have avoided using the magic __hash__ method for this reason. Always
-        use the UID for indexing purposes.
-
-        :return: a hash of all entries in this report group.
-        :rtype: ``int``
-        """
-        return hash(
-            (
-                self.uid,
-                self.status,
-                self.runtime_status,
-                tuple(entry.hash for entry in self.entries),
-            )
-        )
-
-    def xfail(self, strict):
-        """
-        Override report status for test that is marked xfail by user
-        :param strict: whether consider XPASS as failure
-        """
-
-        if self.failed:
-            self.status_override = Status.XFAIL
-        elif self.passed:
-            if strict:
-                self.status_override = Status.XPASS_STRICT
-            else:
-                self.status_override = Status.XPASS
-
-        # do not override if derived status is UNKNOWN/UNSTABLE
-
-        # propagate xfail down to testcase
-        for child in self:
-            child.xfail(strict)
 
 
 class TestReport(BaseReportGroup):
@@ -691,6 +203,22 @@ class TestReport(BaseReportGroup):
         deserialized._index = old_report._index
 
         return deserialized
+
+    def filter(self, *functions, **kwargs) -> Self:
+        """
+        Tag indices are updated after filter operations.
+        """
+        result = super(TestReport, self).filter(*functions, **kwargs)
+
+        # We'd like to call tag propagation before returning the root node,
+        # so we rely on absence of implicit `__copy` arg to decide if we should
+        # trigger tag index propagation or not. If we don't do this check
+        # then tag propagation will be called for each filter call on
+        # sub-nodes which is going to be a redundant operation.
+
+        if kwargs.get("__copy", True):
+            result.propagate_tag_indices()
+        return result
 
 
 class TestGroupReport(BaseReportGroup):
@@ -875,6 +403,22 @@ class TestGroupReport(BaseReportGroup):
             )
         )
 
+    def filter(self, *functions, **kwargs) -> Self:
+        """
+        Tag indices are updated after filter operations.
+        """
+        result = super(TestGroupReport, self).filter(*functions, **kwargs)
+
+        # We'd like to call tag propagation before returning the root node,
+        # so we rely on absence of implicit `__copy` arg to decide if we should
+        # trigger tag index propagation or not. If we don't do this check
+        # then tag propagation will be called for each filter call on
+        # sub-nodes which is going to be a redundant operation.
+
+        if kwargs.get("__copy", True):
+            result.propagate_tag_indices()
+        return result
+
 
 class TestCaseReport(Report):
     """
@@ -888,8 +432,6 @@ class TestCaseReport(Report):
         name,
         tags=None,
         suite_related=False,
-        status_override=None,
-        status_reason=None,
         category=ReportCategories.TESTCASE,
         **kwargs,
     ):
@@ -898,17 +440,8 @@ class TestCaseReport(Report):
         self.tags = tagging.validate_tag_value(tags) if tags else {}
         self.tags_index = copy.deepcopy(self.tags)
         self.suite_related = suite_related
-
-        self.status_override = status_override or Status.NONE
-        self.timer = timing.Timer()
-
         self.attachments = []
-        # testcase is default to passed (e.g no assertion)
-        self._status = Status.UNKNOWN
-        self._runtime_status = RuntimeStatus.READY
         self.category = category
-        self.status_reason = status_reason
-
         self.covered_lines: Optional[dict] = None
 
     def _get_comparison_attrs(self):
