+++ testplan/runners/pools/remote.py
@@ -14,7 +14,6 @@ from testplan.common.remote.remote_resource import (
     RemoteResourceConfig,
     UnboundRemoteResourceConfig,
 )
-from testplan.common.report.base import EventRecorder
 from testplan.common.utils.logger import TESTPLAN_LOGGER
 from testplan.common.utils.path import rebase_path
 from testplan.common.utils.remote import copy_cmd, ssh_cmd
@@ -153,21 +152,15 @@ class RemoteWorker(ProcessWorker, RemoteResource):
 
     def pre_start(self) -> None:
         self.define_runpath()
-        with EventRecorder(
-            name="Prepare remote", event_type="function"
-        ) as event_executor:
+        with self.timer.record("prepare remote"):
             self._prepare_remote()
-        self.event_recorder.add_child(event_executor)
         self._set_child_script()
         self._write_syspath()
 
     def pre_stop(self) -> None:
         """Stop child process worker."""
-        with EventRecorder(
-            name="Fetch results", event_type="function"
-        ) as event_executor:
+        with self.timer.record("fetch results"):
             self._fetch_results()
-        self.event_recorder.add_child(event_executor)
 
     def post_stop(self) -> None:
         self._clean_remote()
