+++ testplan/web_ui/testing/package.json
@@ -100,6 +100,5 @@
     "transformIgnorePatterns": [
       "node_modules/(?!\\.pnpm|rehype-raw|hast-util-raw|unist-util-position|unist-util-visit|unist-util-is|hast-util-from-parse5|hastscript|property-information|hast-util-parse-selector|space-separated-tokens|comma-separated-tokens|vfile-location|web-namespaces|hast-util-to-parse5|zwitch|html-void-elements|axios|chartjs-adapter-date-fns|pretty-bytes).*\\.js"
     ]
-  },
-  "proxy": "http://localhost:59568"
+  }
 }
