+++ tests/unit/testplan/testing/multitest/test_result.py
@@ -22,7 +22,7 @@ from testplan.common.utils import path as path_utils
 from testplan.testing import result as result_mod
 from testplan.testing.multitest import MultiTest
 from testplan.testing.multitest.suite import testcase, testsuite
-from testplan.report.testing import base as report_base
+from testplan.common.report import Status
 
 
 matplotlib.use("agg")
@@ -340,15 +340,15 @@ def test_assertion_skip(mockplan):
     skip_multitest = mtest.report.flatten()[0]
     skip_suite = skip_multitest.entries[0]
     skip_me_case = skip_suite.entries[0]
-    assert skip_me_case.status == report_base.Status.SKIPPED
+    assert skip_me_case.status == Status.SKIPPED
     assert skip_me_case.failed is False
     assert len(skip_me_case.entries) == 2
 
     for index, condition_skip_case in enumerate(skip_suite.entries[1]):
         if index % 2 == 0:
-            assert condition_skip_case.status == report_base.Status.SKIPPED
+            assert condition_skip_case.status == Status.SKIPPED
         else:
-            assert condition_skip_case.status == report_base.Status.PASSED
+            assert condition_skip_case.status == Status.PASSED
         assert len(condition_skip_case.entries) == 1
